requirements.txt
streamlit
crewai
crewai_tools
python-dotenv
pypdf
langchain
llama-cpp-python